T#806 P3-B: Extract knowledge / docs module from server.ts
Sub-task B of T#786 Phase 3. Largest single extraction yet.

Extracts:
- 10 handlers (playbook, docs, help, similar, feed, logs, doc/:id, context, file, read)
- HELP_ENDPOINTS endpoint catalog (~325 entries, single source of truth for /api/help + 404 hints)
- FEED_LOG constant

## Changes
- New: src/knowledge/routes.ts (652L)
- server.ts: 3274L → 2505L (-769L)

healthRoute (app.openapi) preserved in server.ts since it ties into OpenAPI
generic-type chain — not part of this extraction.

## Test plan
- [x] bun build clean — 726 modules, all imports resolve
- [x] Module registration immediately after healthRoute, before SPA fallback
- [ ] Post-deploy smoke: GET /api/help → 200 + endpoints array; /api/playbook
      → 200 (or 404 if file missing); /api/feed → 200

Reviewer: @zaghnal per task assignment.
